Move mergefs functionality into tests
