+++ pkg/cmd/grafana-cli/commands/scuemata_validation_command_test.go
@@ -15,7 +15,7 @@ var defaultBaseLoadPaths = load.GetDefaultLoadPaths()
 func TestValidateScuemataBasics(t *testing.T) {
 	t.Run("Testing scuemata validity with valid cue schemas", func(t *testing.T) {
 		tempDir := os.DirFS(filepath.Join("testdata", "valid_scuemata"))
-		mergedFS := load.Merge(tempDir, defaultBaseLoadPaths.BaseCueFS)
+		mergedFS := Merge(tempDir, defaultBaseLoadPaths.BaseCueFS)
 
 		var baseLoadPaths = load.BaseLoadPaths{
 			BaseCueFS:       mergedFS,
@@ -31,7 +31,7 @@ func TestValidateScuemataBasics(t *testing.T) {
 
 	t.Run("Testing scuemata validity with invalid cue schemas - family missing", func(t *testing.T) {
 		tempDir := os.DirFS(filepath.Join("testdata", "invalid_scuemata_missing_family"))
-		mergedFS := load.Merge(tempDir, defaultBaseLoadPaths.BaseCueFS)
+		mergedFS := Merge(tempDir, defaultBaseLoadPaths.BaseCueFS)
 
 		var baseLoadPaths = load.BaseLoadPaths{
 			BaseCueFS:       mergedFS,
@@ -44,7 +44,7 @@ func TestValidateScuemataBasics(t *testing.T) {
 
 	t.Run("Testing scuemata validity with invalid cue schemas - panel missing", func(t *testing.T) {
 		tempDir := os.DirFS(filepath.Join("testdata", "invalid_scuemata_missing_panel"))
-		mergedFS := load.Merge(tempDir, defaultBaseLoadPaths.BaseCueFS)
+		mergedFS := Merge(tempDir, defaultBaseLoadPaths.BaseCueFS)
 
 		var baseLoadPaths = load.BaseLoadPaths{
 			BaseCueFS:       mergedFS,
